+++ tools/cloc.sh
@@ -18,7 +18,7 @@ SCRIPT_DIR="$( cd "$( dirname "$0" )" && pwd )"
 function countLines() {
   # Run cloc - this counts code lines, blank lines and comment lines for the specified languages.
   # We are only interested in the summary, therefore the tail -1
-  SUMMARY="$(cloc "$1" --include-lang="C++,C/C++ Header,GLSL,JavaScript" --md | tail -1)"
+  SUMMARY="$(cloc "$1" --fullpath --not-match-f="third-party" --include-lang="C++,C/C++ Header,GLSL,JavaScript" --md | tail -1)"
 
   # The $SUMMARY is one line of a markdown table and looks like this:
   # SUM:|101|3123|2238|10783